requirements.txt
pyinstaller
discord.py
discord.py[voice]
discord.py-pagination
discord-py-interactions
interactions-wait-for
pandas
tqdm